+++ petri/incubator/src/lib.rs
@@ -12,6 +12,16 @@
 //!
 //! This crate is backend-agnostic: profiles define the platform requirements,
 //! and incubator backends (currently QEMU TCG) satisfy them.
+//!
+//! # Why QEMU rather than OpenVMM?
+//!
+//! The incubator is fundamentally about providing a *stable host* for running
+//! tests against obscure or emulated hardware (unusual IOMMUs, PCIe topologies,
+//! device-assignment paths, etc.). QEMU TCG is better at faithfully emulating
+//! that breadth of hardware than OpenVMM is, and is likely to remain so. There
+//! is intentionally no OpenVMM backend: testing OpenVMM's own nested-virt
+//! behavior (where the *outer* VMM is under test) is a separate concern handled
+//! elsewhere, not by this crate.
 
 #![forbid(unsafe_code)]
 
@@ -19,6 +29,7 @@ mod profile;
 mod qemu;
 mod run;
 
+pub use profile::Arch;
 pub use profile::IncubatorProfile;
 pub use run::IncubatorConfig;
 pub use run::IncubatorOutput;
