+++ petri/incubator/src/qemu.rs
@@ -6,8 +6,19 @@
 use crate::profile::DeviceConfig;
 use crate::profile::QemuTcgConfig;
 use anyhow::Context;
+use futures::AsyncReadExt;
+use futures_concurrency::future::Race;
+use pal_async::pipe::PolledPipe;
+use pal_async::process::PolledChild;
+use std::collections::BTreeMap;
+use std::io::Write;
 use std::path::Path;
+use std::path::PathBuf;
 use std::process::Command;
+use std::time::Duration;
+
+/// Filename of the injected init script, run by the kernel as `rdinit`.
+const INIT_SCRIPT_NAME: &str = "tcg-init.sh";
 
 /// Build the QEMU command line for a TCG launch.
 pub fn build_qemu_command(
@@ -17,7 +28,6 @@ pub fn build_qemu_command(
     initrd: &Path,
     share_dir: &Path,
     host_pipette_port: u16,
-    kernel_cmdline: &str,
 ) -> anyhow::Result<Command> {
     let mut cmd = Command::new(&config.binary);
 
@@ -28,7 +38,8 @@ pub fn build_qemu_command(
     cmd.arg("-nographic");
     cmd.arg("-kernel").arg(kernel);
     cmd.arg("-initrd").arg(initrd);
-    cmd.arg("-append").arg(kernel_cmdline);
+    cmd.arg("-append")
+        .arg(format!("{} rdinit=/{INIT_SCRIPT_NAME}", config.cmdline));
     cmd.arg("-no-reboot");
 
     // 9p: share the host directory into the guest
@@ -81,7 +92,7 @@ pub fn build_qemu_command(
 /// QEMU's built-in devices use low device numbers. We start at 16 (0x10)
 /// to avoid collisions. The root port for the i-th extra device has
 /// devfn = `(EXTRA_DEVICE_ADDR_BASE + i) << 3`.
-pub const EXTRA_DEVICE_ADDR_BASE: usize = 16;
+const EXTRA_DEVICE_ADDR_BASE: usize = 16;
 
 /// Parse a human-readable size string (e.g., "64M", "1G", "512K") to bytes.
 ///
@@ -114,4 +125,278 @@ fn parse_size(s: &str) -> anyhow::Result<u64> {
         n.checked_mul(multi.unwrap_or(1))
     }()
     .with_context(|| format!("invalid size '{s}'"))
+}
+
+/// Build the guest init script (`rdinit`).
+///
+/// Sets up the environment, mounts the virtio-9p share, brings up networking,
+/// and launches pipette in TCP mode. Pipette then waits for the host to
+/// connect and send commands.
+fn build_init_script() -> String {
+    // QEMU user-mode networking defaults: guest is 10.0.2.15/24,
+    // gateway 10.0.2.2, DNS forwarder at 10.0.2.3.
+    "\
+        #!/bin/sh\n\
+        /bin/busybox --install /bin 2>/dev/null\n\
+        mount -t devtmpfs none /dev\n\
+        mount -t proc none /proc\n\
+        mount -t sysfs none /sys\n\
+        mkdir -p /dev/pts /share /root /tmp /etc\n\
+        mount -t devpts devpts /dev/pts\n\
+        mount -t 9p -o trans=virtio,version=9p2000.L hostshare /share\n\
+        ip link set eth0 up\n\
+        ip addr add 10.0.2.15/24 dev eth0\n\
+        ip route add default via 10.0.2.2\n\
+        echo 'nameserver 10.0.2.3' > /etc/resolv.conf\n\
+        export VMM_TESTS_CONTENT_DIR=/share\n\
+        export HOME=/root\n\
+        cd /share\n\
+        exec /share/pipette --transport tcp\n"
+        .to_string()
+}
+
+/// Prepare the boot initrd by injecting the init script into the base initrd.
+///
+/// Reads the gzip-compressed base initrd, injects the `rdinit` script (see
+/// [`build_init_script`]) under [`INIT_SCRIPT_NAME`], writes the patched initrd
+/// into `share_dir`, and returns its path.
+pub fn prepare_initrd(base_initrd: &Path, share_dir: &Path) -> anyhow::Result<PathBuf> {
+    let initrd_data = std::fs::read(base_initrd).context("failed to read initrd")?;
+
+    let patched_initrd = initrd_cpio::inject_into_initrd(
+        &initrd_data,
+        INIT_SCRIPT_NAME,
+        build_init_script().as_bytes(),
+        0o100755, // regular file, rwxr-xr-x
+    )
+    .context("failed to inject init script into initrd")?;
+
+    let patched_initrd_path = share_dir.join(".incubator-initrd.gz");
+    std::fs::write(&patched_initrd_path, &patched_initrd)
+        .context("failed to write patched initrd")?;
+
+    Ok(patched_initrd_path)
+}
+
+/// Wait for pipette to signal readiness via the serial console relay
+/// task. Races against QEMU exit and a timeout.
+pub async fn wait_for_pipette_ready(
+    driver: &impl pal_async::driver::Driver,
+    timeout: Duration,
+    qemu_child: &mut PolledChild<std::process::Child>,
+    ready_rx: mesh::OneshotReceiver<()>,
+) -> anyhow::Result<()> {
+    enum Event {
+        Ready,
+        QemuExited(std::process::ExitStatus),
+        Timeout,
+    }
+
+    let event = (
+        async {
+            match ready_rx.await {
+                Ok(()) => Event::Ready,
+                // Sender dropped without sending — relay task exited
+                // without seeing the marker (QEMU likely crashed).
+                Err(_) => Event::QemuExited(std::process::ExitStatus::default()),
+            }
+        },
+        async {
+            match qemu_child.wait().await {
+                Ok(status) => Event::QemuExited(status),
+                Err(_) => Event::QemuExited(std::process::ExitStatus::default()),
+            }
+        },
+        async {
+            pal_async::timer::PolledTimer::new(driver)
+                .sleep(timeout)
+                .await;
+            Event::Timeout
+        },
+    )
+        .race()
+        .await;
+
+    match event {
+        Event::Ready => Ok(()),
+        Event::QemuExited(status) => {
+            anyhow::bail!("QEMU exited before pipette was ready (status: {status})");
+        }
+        Event::Timeout => {
+            anyhow::bail!("timed out waiting for pipette ready signal");
+        }
+    }
+}
+
+/// Relay QEMU serial output to a log file, signaling when
+/// pipette's readiness marker appears.
+pub async fn relay_serial_output(
+    mut stdout: PolledPipe,
+    log_path: &Path,
+    ready_tx: mesh::OneshotSender<()>,
+) {
+    let mut log = match std::fs::File::create(log_path) {
+        Ok(f) => f,
+        Err(e) => {
+            tracing::error!(error = %e, "failed to create serial log");
+            return;
+        }
+    };
+
+    let mut ready_tx = Some(ready_tx);
+    let mut buf = vec![0u8; 4096];
+    // Rolling window of recently-seen bytes, used to detect the readiness
+    // marker even if it straddles a read boundary. We never need to retain
+    // more than `marker.len() - 1` bytes between chunks (anything longer
+    // could not be the start of a new match), so this stays small.
+    let marker = pipette_client::PIPETTE_READY_MARKER.as_bytes();
+    let mut window = Vec::new();
+
+    loop {
+        let n = match stdout.read(&mut buf).await {
+            Ok(0) | Err(_) => break,
+            Ok(n) => n,
+        };
+
+        let chunk = &buf[..n];
+        let _ = log.write_all(chunk);
+
+        // Scan for the readiness marker. This does not depend on the marker
+        // being newline-terminated: we append the new bytes and look for the
+        // marker anywhere in the window, then trim the window back down.
+        if ready_tx.is_some() {
+            window.extend_from_slice(chunk);
+            if window.windows(marker.len()).any(|w| w == marker) {
+                if let Some(tx) = ready_tx.take() {
+                    tx.send(());
+                }
+            } else {
+                let keep = marker.len().saturating_sub(1);
+                if window.len() > keep {
+                    window.drain(..window.len() - keep);
+                }
+            }
+        }
+    }
+}
+
+/// Convert a child process's stdout/stderr pipe into a [`std::fs::File`] so it
+/// can be wrapped in a [`PolledPipe`]. The owned-handle type differs by
+/// platform, but the conversion is otherwise identical.
+#[cfg(unix)]
+pub fn child_pipe_to_file(pipe: impl Into<std::os::unix::io::OwnedFd>) -> std::fs::File {
+    std::fs::File::from(pipe.into())
+}
+
+#[cfg(windows)]
+pub fn child_pipe_to_file(pipe: impl Into<std::os::windows::io::OwnedHandle>) -> std::fs::File {
+    std::fs::File::from(pipe.into())
+}
+
+/// Set up VFIO devices inside the incubator.
+///
+/// Each extra device in the profile sits behind its own PCIe root port
+/// at a known PCI device number (see [`EXTRA_DEVICE_ADDR_BASE`]). This
+/// function discovers the child device's BDF by finding the bridge at
+/// that slot in sysfs, then unbinds the child from its driver and binds
+/// it to vfio-pci.
+///
+/// Returns a map of environment variables to set for the guest command,
+/// e.g., `INCUBATOR_VFIO_BDF_TEST_DISK=0000:01:00.0`.
+pub async fn setup_vfio_devices(
+    client: &pipette_client::PipetteClient,
+    devices: &[DeviceConfig],
+) -> anyhow::Result<BTreeMap<String, String>> {
+    let mut env = BTreeMap::new();
+
+    // Collect (device_index, config) for devices that need VFIO binding.
+    let vfio_devices: Vec<_> = devices
+        .iter()
+        .enumerate()
+        .filter_map(|(i, d)| match d {
+            DeviceConfig::VirtioBlk(cfg) if cfg.vfio => Some((i, cfg)),
+            DeviceConfig::VirtioBlk(_) => None,
+        })
+        .collect();
+
+    if vfio_devices.is_empty() {
+        return Ok(env);
+    }
+
+    tracing::info!("setting up {} VFIO device(s)", vfio_devices.len());
+
+    for (device_index, cfg) in &vfio_devices {
+        let addr = EXTRA_DEVICE_ADDR_BASE + device_index;
+
+        // The root port for this device is deterministically at
+        // 0000:00:{addr:02x}.0 (see `build_qemu_command`). Read its
+        // secondary bus number from sysfs; the assigned device sits at
+        // slot 0, function 0 of that bus.
+        let rp_bdf = format!("0000:00:{addr:02x}.0");
+        let secondary_bus_path = format!("/sys/bus/pci/devices/{rp_bdf}/secondary_bus_number");
+        let secondary_bus_raw = client
+            .read_file(&secondary_bus_path)
+            .await
+            .with_context(|| {
+                format!(
+                    "failed to read secondary bus number for device '{}' (root port {rp_bdf})",
+                    cfg.name
+                )
+            })?;
+        // sysfs reports the secondary bus number in decimal.
+        let secondary_bus_str = String::from_utf8_lossy(&secondary_bus_raw);
+        let secondary_bus: u8 = secondary_bus_str.trim().parse().with_context(|| {
+            format!(
+                "unexpected secondary bus number {secondary_bus_str:?} for device '{}'",
+                cfg.name
+            )
+        })?;
+        let bdf = format!("0000:{secondary_bus:02x}:00.0");
+
+        // Confirm the child device actually exists before trying to rebind it.
+        client
+            .read_file(format!("/sys/bus/pci/devices/{bdf}/vendor"))
+            .await
+            .with_context(|| {
+                format!(
+                    "no device found behind root port {rp_bdf} (expected {bdf}) for device '{}'",
+                    cfg.name
+                )
+            })?;
+
+        tracing::info!(name = %cfg.name, %bdf, %addr, "binding device to vfio-pci");
+
+        // Unbind from current driver
+        let _ = client
+            .write_file(
+                format!("/sys/bus/pci/devices/{bdf}/driver/unbind"),
+                bdf.as_bytes(),
+            )
+            .await;
+
+        // Set driver override to vfio-pci
+        client
+            .write_file(
+                format!("/sys/bus/pci/devices/{bdf}/driver_override"),
+                b"vfio-pci".as_slice(),
+            )
+            .await
+            .context("failed to set driver_override")?;
+
+        // Bind to vfio-pci
+        client
+            .write_file("/sys/bus/pci/drivers/vfio-pci/bind", bdf.as_bytes())
+            .await
+            .context("failed to bind to vfio-pci")?;
+
+        // Export env var: name "test-disk" → INCUBATOR_VFIO_BDF_TEST_DISK
+        let env_name = format!(
+            "INCUBATOR_VFIO_BDF_{}",
+            cfg.name.to_uppercase().replace('-', "_")
+        );
+        tracing::info!(%env_name, %bdf, "VFIO device ready");
+        env.insert(env_name, bdf);
+    }
+
+    Ok(env)
 }
